refactor: rename IndexedStrategy Utils lib to SXUtils lib

+++ src/utils/BitPacker.sol
@@ -2,10 +2,11 @@
 
 pragma solidity ^0.8.18;
 
-// @dev Append only bit packing library
+/// @title Library for setting and checking bits in a uint256.
 library BitPacker {
+    /// @dev Sets the bit at the given index to the given value.
     function setBit(uint256 value, uint8 index, bool bit) internal pure returns (uint256) {
-        uint256 mask = uint256(1 << index);
+        uint256 mask = 1 << index;
         if (bit) {
             return value | mask;
         } else {
@@ -13,8 +14,9 @@ library BitPacker {
         }
     }
 
+    /// @dev Returns true if the bit at the given index is set.
     function isBitSet(uint256 value, uint8 index) internal pure returns (bool) {
-        uint256 mask = uint256(1 << index);
+        uint256 mask = 1 << index;
         return (value & mask) != 0;
     }
 }
